re-org package and define public api
config docs
