+++ cdippy/plots/__init__.py
@@ -1,0 +1,3 @@
+from . import annual_hs_boxplot, compendium, sst_climatology
+
+__all__ = ["annual_hs_boxplot", "compendium", "sst_climatology"]
