+++ docs/api.md
@@ -1,0 +1,18 @@
+# CDIPpy API reference
+
+::: cdippy.cdipnc
+
+
+::: cdippy.nchashes
+
+
+::: cdippy.ncstats
+
+
+::: cdippy.ndbc
+
+
+::: cdippy.spectra
+
+
+::: cdippy.stndata
